Writing clean JPEG files now, after removing static int
In the writeToSD function there was a variable declaration static int i
= 0; This appears to have been preserving the i value between function
calls (separate image-to-SD write events), and so adding some of the
previous image's data to the start of the current image, making a
semi-corrupted jpeg. Changing i to a standard int instead of static int
appears to have fixed this.